# plugins/cmd/buildpermission-plugin/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Azure DevOps BuildPermission Plugin API for Krateo Operator Generator (KOG)
  description: Simple wrapper around Azure DevOps API to provide consistency of API response for Krateo Operator Generator (KOG)
  termsOfService: http://swagger.io/terms/
  contact:
    name: Krateo Support
    url: https://krateo.io
    email: contact@krateoplatformops.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
servers:
  - url: http://localhost:8080/
paths:
  /plugin/buildpermission/{organization}/{projectId}:
    get:
      tags:
        - Build Permissions
      summary: Get build permissions
      description: Retrieve the permissions for a specific build definition in Azure DevOps
      operationId: get-buildpermission
      parameters:
        - name: organization
          in: path
          description: Organization name
          required: true
          schema:
            type: string
        - name: projectId
          in: path
          description: Project ID
          required: true
          schema:
            type: string
        - name: buildDefinitionId
          in: query
          description: Build Definition ID
          schema:
            type: string
        - name: identityType
          in: query
          description: Type of identity (e.g., 'azure-group', 'build-service')
          schema:
            type: string
        - name: identityName
          in: query
          description: Name of the identity (e.g., Contributors), not used if identityType is 'build-service' or if identityDescriptor is provided
          schema:
            type: string
        - name: identityDescriptor
          in: query
          description: Descriptor of the identity, has priority over identityType and identityName
          schema:
            type: string
        - name: projectLevel
          in: query
          description: Whether to manage permissions at the project level (true) or build level (pipeline level) (false). Default is false.
          schema:
            type: boolean
            default: false
        - name: api-version
          in: query
          description: API version (e.g., 7.2-preview.2)
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Basic Auth header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Retrieved build permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/buildpermission.BuildPermissionResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
    post:
      tags:
        - Build Permissions
      summary: Set build permissions
      description: Set or update the permissions for a specific build definition in Azure DevOps. Identity information comes from the request body.
      operationId: post-buildpermission
      parameters:
        - name: organization
          in: path
          description: Organization name
          required: true
          schema:
            type: string
        - name: projectId
          in: path
          description: Project ID
          required: true
          schema:
            type: string
        - name: buildDefinitionId
          in: query
          description: Build Definition ID
          schema:
            type: string
        - name: projectLevel
          in: query
          description: Whether to manage permissions at the project level (true) or build level (pipeline level) (false). Default is false.
          schema:
            type: boolean
            default: false
        - name: api-version
          in: query
          description: API version (e.g., 7.2-preview.2)
          required: true
          schema:
            type: string
        - name: Authorization
          in: header
          description: Basic Auth header
          required: true
          schema:
            type: string
      requestBody:
        description: Permission settings to apply. Identity info (type, name, descriptor) is specified in permissions.identity.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/buildpermission.BuildPermissionRequest'
        required: true
      responses:
        "200":
          description: Updated build permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/buildpermission.BuildPermissionResponse'
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: string
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: string
      x-codegen-request-body-name: body
components:
  schemas:
    buildpermission.BuildPermissionIdentity:
      type: object
      properties:
        descriptor:
          type: string
        name:
          type: string
        type:
          type: string
    buildpermission.BuildPermissionRequest:
      type: object
      properties:
        permissions:
          $ref: '#/components/schemas/buildpermission.PermissionsInfo'
    buildpermission.BuildPermissionResponse:
      type: object
      properties:
        buildDefinitionId:
          type: string
        organization:
          type: string
        permissions:
          $ref: '#/components/schemas/buildpermission.PermissionsInfo'
        projectId:
          type: string
        projectLevel:
          type: boolean
    buildpermission.PermissionsInfo:
      type: object
      properties:
        allow:
          type: object
          additionalProperties:
            type: boolean
        deny:
          type: object
          additionalProperties:
            type: boolean
        identity:
          $ref: '#/components/schemas/buildpermission.BuildPermissionIdentity'
x-original-swagger-version: "2.0"